+++ tests/json_data.sql
@@ -1,0 +1,3 @@
+select 'json' as component;
+select 'It works!' as message;
+select 'cool' as cool;